requirements.txt
django
djangorestframework
Pillow
djangorestframework-camel-case
drf-yasg